pyproject.toml
[tool.poetry]
name = "github-metrics"
version = "0.1.0"
description = ""
authors = ["Jakub Valenta <jakub@jakubvalenta.cz>"]
license = "Apache Software License"
readme = "README.md"
packages = [{include = "github_metrics"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
mypy = "^0.982"
flake8 = "^5.0.4"
isort = "^5.10.1"
types-requests = "^2.28.11.2"
types-python-dateutil = "^2.8.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
